fix: ensure timeout set by RetryMiddleware is int not float (https://github.com/googleapis/gax-php/issues/462)

+++ Gax/src/Middleware/RetryMiddleware.php
@@ -128,7 +128,7 @@ class RetryMiddleware
         }
 
         $delayMs = min($delayMs * $delayMult, $maxDelayMs);
-        $timeoutMs = min(
+        $timeoutMs = (int) min(
             $timeoutMs * $timeoutMult,
             $maxTimeoutMs,
             $deadlineMs - $this->getCurrentTimeMs()
